Simplify club manage btn
